fix(test): make sign-compare GCC-clean and correct DatabaseWorker test counts
CI 门禁 PR #16 四个 Linux job 失败：test_database_worker.cpp:401
comparison of integer expressions of different signedness（uint64_t
failed_count() vs int cancelled，-Werror=sign-compare）。MSVC 不启用该警告
故本地未拦截。修复：比较处显式 static_cast<std::uint64_t>。

顺带修正文档事实：test_database_worker 实测 8 test case / 102 断言
（debug 与 release 双配置 Catch2 实测一致），里程碑与总计划三处误记 7/87、
一处稳定性次数误记 40 次，按本 PR 实测（debug 60 次/release 30 次）修正并
注明（沿用 M1-05/M2-04 计数修正先例）。

本地验证：GCC 15.2.0 -Wall -Wextra -Wpedantic -Werror -fsyntax-only 对
test_database_worker.cpp/database_worker.cpp/database.cpp 全部通过（含
CI 同款警告集）；MSVC debug 重编译零告警、ctest 16/16、worker 测试 8/102
全过；release 16/16。

+++ tests/unit/test_database_worker.cpp
@@ -398,7 +398,7 @@ TEST_CASE("Drain budget exhaustion is recorded and not faked as clean",
         }
     }
     REQUIRE(executed.load() + cancelled == 3);
-    REQUIRE(fx.control->failed_count() == cancelled);
+    REQUIRE(fx.control->failed_count() == static_cast<std::uint64_t>(cancelled));
 
     REQUIRE(report.executor_shutdown_completed);
 }
